refactor(teams): split orchestrator.rs into focused submodules [OPE-402]
Split the monolithic orchestrator.rs (573 LOC) into four focused modules:
- mod.rs: TeamOrchestrator struct definition and construction
- dispatch.rs: execute() and resume() workflow dispatch methods
- delegation.rs: process_pending_delegations() recursive delegation handler
- helpers.rs: shared pub(crate) utilities (process_agent_communications, etc.)
- tests.rs: all 6 original tests + 4 new edge-case tests

Public API unchanged. All 274 tests pass (257 unit + 17 integration).
